NSXv3: Search for the default section from the end
Upon startup we inquire the backend for the list of firewall section, as
we want to find the specific default section.
Because the firewall section should be located at the bottom, it is more
likely it is located at the end of the list, thus iterating the list in
reversed fashion may save few cycles on startup.

Change-Id: I6212adaa4924cbd46b9d810e95f737d341e10883