Add Constraints support
Adding constraints support to libraries is slightly more complex than
services as the libraries themselves are listed in upper-constraints.txt
which leads to errors that you can't install a specific version and a
constrained version.

This change adds constraints support by also adding a helper script to
edit the constraints to remove vmware-nsxlib.

Change-Id: Ibbb6da2551bf7eb840fb4b1d11e090dbe4db7c52