Fix address bindings in logical port update
During port updates, if the user does not pass address bindings,
nsxlib should perform a LP GET on the backend and use the
existing address bindings.
The response body returns address bindings in a dict format
which breaks the update in _build_attrs method. Instead of a
dot lookup on the attributes, we shall use .get() lookup for
address bindings which should work on both dict types and
objects.

Change-Id: I660cc63264d1458d17d587555889974571960bd5