Allow passing args of type list for NSGroup and firewall rule methods
The current NSGroup create method does not allow for passing a
list of membership criterias. Similarly the source, destination
and service arguments of Firewall rule method does not allow for
passing a list. This patch provides a fix for it
while maintaining backwards compatibility where in the caller may
pass a single argument instead of a list.

Change-Id: I6b16d37bf3ca61f3c9f02688f9548ea4b3b6adb6