Enable external context managers for client certificates
The vmware_nsxlib.v3.cluster.Provider class now allows for specifying
a client certificate context manager. This is supposed to be an
user-supplied context manager object with a cert_file attribute that
will be associated with the NSX session. The operations performed
in the __enter__ and __exit__ methods pertain domain specific logic
but are required to ensure the cert_file attribute is defined within
the context manager.

If the Provider object specifies a context manager for client
certificate authentication, it will take priority over the cert_file
attribute in the same class.

Change-Id: Id62b299ed56d275e3a9d8db4eaf302cd40d49a5f