NSXv3: Fix init of default firewall section
Default firewall section for neutron is created once and serves all
openstack deployments on same backend. Rules for this section are
updated on each neutron init.
This section is not deleted with devstack unstack. If the section was
created with client auth (=in protected mode), next stack with same
backend may fail due to identity conflict.
This change forces the update, assuming current user is a superuser.

Change-Id: I0ad20fb15d760d56a792e9fd94fcc0efcc38449a