Extend client silent mode for logical switch
Allow resources get operations to be silent, and also not log
the validate result warning if silent.
The reason is that get actions are sometimes used in order to verify
that the object does not exist, and so we do not want to log it.

(Cherry picked from: Ib32637da86e72ff22a7c5684a3f179b91f09406f)

Change-Id: Ic4420d04957058cd0a182da23816e3667060e865