Support bypass-firewall parma for router NAT rules
Adding nat_pass parameter to router NAT rules creation.
The backend default value for this parameter is True, meaning the
firewall rules will be bypassed.
So this is also the default of the api, for backwards compatibility.

This parameter is acceptable only since nsx v2, so the api needs the
ability to check the supported features before creting the request body.
For that use, the core_resources will now get the nsxlib object at init.

Change-Id: I7c9dfe13a146a47b180fc3df2d4d6174f252e0a3