[WIP] create session for XSRF token
To avoid CSRF attacks, the requests to the NSX manager should have
X-XSRF-TOKEN header.
To get it, and the JSESSIONID, the cluster will issue a session/create
request for each endpoint at init, and retrieve the values from the response
header.
Those values will later be used for each request to this endpoint.

Change-Id: I24cf6416e38dc7f57d7d8ceece48a1d3d5815112