Write and delete client cert for each request
Current version of requests library checks for client certificate
file existance on each request, regardless of whether the cert is
going to be needed for this request.
Therefore it is useless to save the effort of populating cert file
for non-first requests on the connection.

In addition, add a retry for the case SSL error comes from within
SSL C code.

(cherry picked from commit 187651405b5b537c584d7658cb3524bf2081f99a)

Change-Id: Ie3710f017f6109d28b6cf956de020465552d4b3a