Write and delete client cert for each request
Current version of requests library checks for client certificate
file existance on each request, regardless of whether the cert is
going to be needed for this request.
Therefore it is useless to save the effort of populating cert file
for non-first requests on the connection.

In addition, add a retry for the case SSL error comes from within
SSL C code.

Change-Id: I32b8304b3217049752e8d25a1b735a6d2035fa0b