Provide a callback to inject headers to the NSX
Enable application to bind to a callback that will inject
headers to the requests sent to the NSX. This can for example
be: 'X-NSX-EUSER' which will provide the NSX context of the user
on behalf of whom the opertaion is done.

A new method set_inject_headers_callback(in is added. This method
receives a callback that should return a dictionary of headers to
be added.

Change-Id: I90fa7ea8c7828bf97aec66321e4169588108760e
(cherry picked from commit 216d8a3be98e17a664afd6a40a98667c3ccc1f0c)