Allow creating firewall section with empty rule
Currently if users pass rules=[] when calling
create_with_rules method, it will get errors from MP.
The reason is that this method does not append
create_with_rules action in the url if rules is None or empty,
but it still includes 'rules=[]' attribute in the request body.

Thus MP throws an error because it doesn't expect 'rules'
attribute in the request.

Change-Id: I86122f1f2b5382224084d786f9700330e2c83d50