NSX rate limit support
In case of too many requests in a short period of time, the NSX will
return response 429.
In this case (if configured) the nsxlib client will retry sending the request.

This option is controlled by a new parameter in the nsxlib config
rate_limit_retry which is enabled by default.

Change-Id: I20fca36d553e1e74da61292342a87247b53b5d13