fix tox python3 overrides
We want to default to running all tox environments under python 3, so
set the basepython value in each environment.

We do not want to specify a minor version number, because we do not
want to have to update the file every time we upgrade python.

We do not want to set the override once in testenv, because that
breaks the more specific versions used in default environments like
py35 and py36.

Depends-on: I660a4b87763fdb67ef7be7419536449a90568013
Change-Id: I8b61fcf246647ea031d05e3604c889e29b92beb3