NSXP LB: Remove monitor only if pool has it
The code attempts to remove a monitor from the pool even if the pool doesn't
use the monitor.
Although this condition is invalid, it may occur due to inconsistencies
between Neutron and the backend.
When such inconsistency occurs, we could raise an exception. But, as this is
a delete operation, and the users wants this to be gone - if it's not there,
then we are good...

Change-Id: Ifce56be6846b85130ea716c95a11be4d33db177b